chore: Brand file 분리
